requirements.txt
flask
torch
torchvision
open_clip_torch
pandas
scikit-learn
Pillow
